pytorch-models/requirements.txt
gensim
pyspellchecker
numpy